// src/types.ts
interface Condition {
  type: 'address' | 'bytes' | 'string' | 'bignumber';
  condition: string;
  value: string;
  index: number;
  valueFormatted?: any;
}

export interface ConditionStatus {
  conditions: Condition[];
  isComplete: boolean;
}

export interface TransactionDescriptionAndConditionsStatus {
  id: string;
  target: string;
  hash: string;
  parsed: TransactionDescription;
  conditions: Condition[];
  isComplete: boolean;
  address: string;
}